Fix clippy errors in slab-whisper: unnecessary cast and Arc Send/Sync

+++ slab-whisper/src/whisper_state/mod.rs
@@ -224,7 +224,7 @@ impl WhisperState {
         if ret < 0 {
             Err(WhisperError::GenericError(ret))
         } else {
-            Ok((ret as i32, lang_probs))
+            Ok((ret, lang_probs))
         }
     }
